Fixing hunger and thirst INI props
git-svn-id: svn://192.168.1.10/public/CoffeeMud@23494 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Abilities/Properties/Prop_UseAdjuster.java
@@ -126,10 +126,10 @@ public class Prop_UseAdjuster extends Prop_Adjuster implements ArchonOnly
 					charState.adjHitPoints((int)Math.round(CMath.mul(charState.getHitPoints(), CMath.div(((Integer) changes[c + 1]).intValue(),100))-charState.getHitPoints()),mob.maxState());
 					break;
 				case CharState.STAT_HUNGER:
-					charState.adjHunger(((Integer) changes[c + 1]).intValue(),mob.maxState().maxHunger(mob.baseWeight()));
+					charState.adjHunger(((Integer) changes[c + 1]).doubleValue(),mob.maxState().maxHunger(mob.baseWeight()));
 					break;
 				case CharState.STAT_THIRST:
-					charState.adjThirst(((Integer) changes[c + 1]).intValue(),mob.maxState().maxThirst(mob.baseWeight()));
+					charState.adjThirst(((Integer) changes[c + 1]).doubleValue(),mob.maxState().maxThirst(mob.baseWeight()));
 					break;
 				case CharState.STAT_MANA:
 					charState.adjMana((int)Math.round(CMath.mul(charState.getMana(), CMath.div(((Integer) changes[c + 1]).intValue(),100))-charState.getMana()),mob.maxState());
